Implement rating related services and repositories and calculation logic

+++ src/main/java/com/karifovas/gamerating/model/RatingConfiguration.java
@@ -1,5 +1,6 @@
 package com.karifovas.gamerating.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record RatingConfiguration(
@@ -7,8 +8,8 @@ public record RatingConfiguration(
         List<Rating> ratings
 ) {
     public record Scale(
-            float min,
-            float max
+            BigDecimal min,
+            BigDecimal max
     ) {
     }
 }
